Refactored windows permission handling into a separate file
